refactor(flight): decompose server.py (1730 to 557 lines)
Phase 5.5 (final Phase 5 decomposition): split the OBFlightServer
gravity-well into a facade plus three sibling modules. Pure code movement;
zero behavior change. OBFlightServer keeps the Flight overrides
(get_flight_info, do_get, do_action, list_flights), __init__, and pending-
ticket helpers; every extracted helper is a one-line delegator.

- ob_flight/server_routing.py: session/model routing (_SessionRouting*
  middleware, selector/model resolution, error formatters)
- ob_flight/server_catalog.py: catalog/metadata (catalog tables/columns,
  virtual-table detection, schema probing, model->catalog builders)
- ob_flight/server_execution.py: SQL classify/rewrite/prepare/execute +
  cache get/put + arrow type hints + mode constants

New modules type-hint the server under TYPE_CHECKING and never import
ob_flight.server -> 0 cycles. server.py re-exports the routing classes,
mode constants, and db_connect (tests patch ob_flight.server.db_connect).

Driver suite 150 passed; driver ruff + mypy clean; main repo unaffected
(2487 passed).
